+++ lib/src/subtypes/kingdoms/kingdom_type.dart
@@ -21,9 +21,6 @@ abstract class KingdomType {
   /// Returns a generator for the type of other important settlements in the kingdom
   IGenerator<SettlementType> getImportantSettlementsTypesGenerator();
 
-  /// Returns a generator for the government type of the kingdom
-  IGenerator<GovernmentType> getGovernmentTypeGenerator();
-
   /// Returns a generator for something a kingdom is known for
   IGenerator<String> getKnownForGenerator();