+++ CHANGELOG.md
@@ -24,6 +24,8 @@
 
 - Fixed some of the descriptions of some entities
 
+- Added government type as an argument for kingdom generator
+
 ## 0.8.1
 
 - Changed descriptor generator to be based on alignment